+++ hsolve/HSolveActive.h
@@ -139,7 +139,7 @@ protected:
     vector<ChannelData>		  channel_data_;
     ChannelData 			  * channel_data_d;
     void resetDevice();
-    void copy_to_device(float ** v_row_array, float * v_row_temp, int size);
+    void copy_to_device(double ** v_row_array, double * v_row_temp, int size);
     LookupColumn			  *column_d;
     int                       is_inited_;
 	void copy_data(std::vector<LookupColumn>& column,
@@ -191,8 +191,8 @@ private:
 
 #ifdef USE_CUDA
 	void advanceChannel_gpu(
-    float *                          v_row,
-    vector<float>&                   caRow,
+    double *                          v_row,
+    vector<double>&                   caRow,
     LookupColumn                    * column,                                           
     LookupTable&                     vTable,
     LookupTable&                     caTable,
